requirements.txt
Authlib==1.3.0
Bootstrap-Flask==2.2.0
email_validator==2.1.1
Flask==2.3.2
Flask-OAuthlib==0.9.6
Flask-WTF==1.2.1
Jinja2==3.1.3
oauthlib==2.1.0
waitress==3.0.0
psycopg2_binary==2.9.9
Flask-Bootstrap==3.3.7.1
pytz==2024.1
WTForms==3.0.1
email_validator==2.1.1
flask-babel==4.0.0
flask-login==0.6.3
SQLAlchemy==2.0.29
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1